Made plugin more modular
