+++ ocr_translate_hugging_face/plugin/__init__.py
@@ -1,0 +1,26 @@
+###################################################################################
+# ocr_translate-hugging_face - a plugin for ocr_translate                         #
+# Copyright (C) 2023-present Davide Grassano                                      #
+#                                                                                 #
+# This program is free software: you can redistribute it and/or modify            #
+# it under the terms of the GNU General Public License as published by            #
+# the Free Software Foundation, either version 3 of the License.                  #
+#                                                                                 #
+# This program is distributed in the hope that it will be useful,                 #
+# but WITHOUT ANY WARRANTY; without even the implied warranty of                  #
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the                   #
+# GNU General Public License for more details.                                    #
+#                                                                                 #
+# You should have received a copy of the GNU General Public License               #
+# along with this program.  If not, see {http://www.gnu.org/licenses/}.           #
+#                                                                                 #
+# Home: https://github.com/Crivella/ocr_translate-hugging_face                    #
+###################################################################################
+"""Collection of importable plugins."""
+from .seq2seq import HugginfaceSeq2SeqModel
+from .ved import HugginfaceVEDModel
+
+__all__ = (
+    'HugginfaceSeq2SeqModel',
+    'HugginfaceVEDModel',
+)
